+++ examples/loans/CHANGELOG.md
@@ -1,5 +1,20 @@
 # xchainjs-loans
 
+## 1.0.8
+
+### Patch Changes
+
+- @xchainjs/xchain-thorchain-amm@1.1.7
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [b379aeb]
+- Updated dependencies [b379aeb]
+  - @xchainjs/xchain-cosmos@1.1.0
+  - @xchainjs/xchain-thorchain-amm@1.1.6
+
 ## 1.0.6
 
 ### Patch Changes